+++ assets/src/scripts/d3-rendering/tooltips.spec.js
@@ -2,7 +2,6 @@ import {select} from 'd3-selection';
 
 import {appendTooltip} from './tooltips';
 
-
 describe('d3-rendering/tooltips', () => {
 
     let testDiv;